requirements.txt
aiohttp==2.3.1
aiohttp-jinja2==0.14.0
aiohttp-session==1.0.1
aiopg==0.13.1
aioredis==0.3.4
asn1crypto==0.23.0
async-timeout==2.0.0
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.2
hiredis==0.2.0
idna==2.6
Jinja2==2.9.6
MarkupSafe==1.0
multidict==3.3.0
peewee==2.10.2
peewee-async==0.5.7
pkg-resources==0.0.0
psycopg2==2.7.3.2
pycparser==2.18
python-engineio==1.7.0
python-socketio==1.8.1
redis==2.10.6
six==1.11.0
yarl==0.13.0
